Update TypeScript and package.json configurations
