+++ tsconfig.json
@@ -1,10 +1,17 @@
 {
   "compilerOptions": {
     "strict": true,
+    "target": "ESNext",
+    "module": "CommonJS",
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
-    "module": "ESNext",
-    "target": "ESNext",
-    "moduleResolution": "node",
-  }
+    "skipLibCheck": true,
+    "outDir": "./dist",
+  },
+  "exclude": [
+    "node_modules",
+  ],
+  "include": [
+    "src"
+  ]
 }
